requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
opencv-python-headless
ultralytics
Pillow
numpy
